+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 ### master
 
+### 1.3.4 (13-10-2021)
+
+#### Tooltip
+
+- Properly render serie name by escaping html
+
+#### Legend
+
+- Fix a problem with quoted series names
+
+### 1.3.3 (11-10-2021)
+
 #### Tooltip
 
 -   Store tooltip plugin in `yagr.plugins` for direct access to methods
 -   Unpin and hide tooltip when plot resizing.
+
+#### Processing
+
+-   Added `snapToValues = false` for interpolation cursor snapping